Implementing the formatting class
adds a button to IO
